♻ Extremely large refactor to improve configure+build times
This has also broken up the main Includes/Project.cmake file into
several.

🔥 Removed FindDoxygen.cmake (We will not be using Doxygen)
🔥 Remove FindIWYU.cmake (include-what-you-use is error prone and hates
our codebase)
✨ Add add_sphinx_target to generate a docs target
✨ Add a bunch of properties to help sphinx generate even better
documentation
✨ Add top-level docs target that can have others plugged into it

This allows building multiple projects, however, it does not currently
act like rustdoc and some work is needed to create "amalgamate"
documentation so that everything is ready and present.

🐛 Fixed "regeneration stall" in CMake

Sometimes allowing CMake to reconfigure itself on a project would stall
the project for up to an hour. The only solution to fix this was to kill
the build, and manually reconfigure the project. This was fixed in a
patch in CMake 3.16, however the version on ubuntu does not currently
feature this patch.

🚨 Add several warnings to C++ tests
⬆ Upgrade Catch2 from v2.12.1 to v2.13.2