CMake Future Proofing Update (#2)
✨ Add add_sphinx_target to generate a docs target
✨ Add a bunch of properties to help sphinx generate even better documentation
✨ Add top-level docs target that can have others plugged into it

This allows building multiple projects, however, it does not currently act like rustdoc and some work is needed to create "amalgamate" documentation so that everything is ready and present.

⬆ Upgrade Catch2 from v2.12.1 to v2.13.2

🔨 Add several warning flag checks for clang-11
🔨 netlify-tests target is now EXCLUDE_FROM_ALL
🔨 Add SPHINX_PROJECT_AUTHOR property

🐛 Fixed "regeneration stall" in CMake

Sometimes allowing CMake to reconfigure itself on a project would stall
the project for up to an hour. The only solution to fix this was to kill
the build, and manually reconfigure the project. This was fixed in a
patch in CMake 3.16, however the version on ubuntu does not currently
feature this patch.

🚨 Add several warnings to C++ tests

📝 Add config settings for sphinx

💄 Add div.note to custom.css

🚚 Rename conf.py.in to conf.py

♻ Extremely large refactor to improve configure+build times
♻ Add SPHINX_PROJECT_NAME property for document properties
♻ Refactor generating docs target
We now have a bunch of custom settings to make the conf.py work much
better and the generated CSS is way nicer to look at.

🔥 Remove superfluous Includes/Project.cmake
🔥 Removed FindDoxygen.cmake (We will not be using Doxygen)
🔥 Remove FindIWYU.cmake 
include-what-you-use is error prone and hates our codebase